Moved repeated elements in contructor to superclass
