+++ Car.java
@@ -1,9 +1,18 @@
+import java.awt.*;
+
 public abstract class Car {
+    private int nrDoors;
+    private Color color;
+    protected int enginePower;
+    private String modelName;
+    private double currentSpeed;
+    
     public Car(int nrDoors, Color color, int enginePower, String modelName){
         this.nrDoors = nrDoors;
         this.color = color;
         this.enginePower = enginePower;
         this.modelName = modelName;
+        stopEngine();
     }
     
     public int getNrDoors(){
@@ -43,6 +52,15 @@ public abstract class Car {
     public void brake(double amount){
         decrementSpeed(amount);
     }
+
+    public void incrementSpeed(double amount){
+        currentSpeed = getCurrentSpeed() + speedFactor() * amount;
+    }
+
+    public void decrementSpeed(double amount){
+        currentSpeed = getCurrentSpeed() - speedFactor() * amount;
+    }
+    
 
     // Jag testar om jag har write access
 }
