+++ Volvo240.java
@@ -3,31 +3,15 @@ import java.awt.*;
 public class Volvo240 extends Car{
 
     public final static double trimFactor = 1.25;
-    public int nrDoors; // Number of doors on the car
-    public double enginePower; // Engine power of the car
-    public double currentSpeed; // The current speed of the car
-    public Color color; // Color of the car
-    public String modelName; // The car model name
     
-    public Volvo240(){
-        nrDoors = 4;
-        color = Color.black;
-        enginePower = 100;
-        modelName = "Volvo240";
-        stopEngine();
+    public Volvo240(Color color){
+        super(4, color, 100, "Saab240");
     }
     
     public double speedFactor(){
         return enginePower * 0.01 * trimFactor;
     }
 
-    public void incrementSpeed(double amount){
-	    currentSpeed = Math.min(getCurrentSpeed() + speedFactor() * amount,enginePower);
-    }
-
-    public void decrementSpeed(double amount){
-        currentSpeed = Math.max(getCurrentSpeed() - speedFactor() * amount,0);
-    }
-
+   
 
 }
